Quick refactor of mappage since it was getting out of hand
