+++ src/components/FeatureCounter.tsx
@@ -1,20 +1,13 @@
-// src/components/FeatureCounter.tsx
 import React from 'react';
 
 interface FeatureCounterProps {
-  featureCounts: number[];
+  count: number;
 }
 
-const FeatureCounter: React.FC<FeatureCounterProps> = ({ featureCounts }) => {
-  return (
-    <div className="absolute bottom-4 right-4">
-      {featureCounts.map((count, index) => (
-        <div key={index}>
-          Layer {index + 1}: {count} features visible
-        </div>
-      ))}
-    </div>
-  );
-};
+const FeatureCounter: React.FC<FeatureCounterProps> = ({ count }) => (
+  <div className="absolute bottom-4 right-4 bg-white p-2 rounded shadow">
+    Visible Features: {count}
+  </div>
+);
 
 export default FeatureCounter;
